+++ library/src/lapack/roclapack_geqrf.hpp
@@ -97,8 +97,6 @@ void rocsolver_geqrf_getMemorySize(const I m,
     }
 }
 
-#define ROCSOLVER_GEQRF_USE_HIPGRAPH
-
 template <bool BATCHED, bool STRIDED, typename T, typename I, typename U>
 rocblas_status rocsolver_geqrf_template(rocblas_handle handle,
                                         const I m,
@@ -141,15 +139,19 @@ rocblas_status rocsolver_geqrf_template(rocblas_handle handle,
     I ldw = GEQxF_BLOCKSIZE;
     rocblas_stride strideW = rocblas_stride(ldw) * ldw;
 
-#ifdef ROCSOLVER_GEQRF_USE_HIPGRAPH
+    // enable graph capture
+    rocsolver_alg_mode alg_mode;
+    ROCBLAS_CHECK(rocsolver_get_alg_mode(handle, rocsolver_function_geqrf, &alg_mode));
 
     hipStream_t graph_stream;
-    HIP_CHECK(hipStreamCreate(&graph_stream));
-    rocblas_set_stream(handle, graph_stream);
-    HIP_CHECK(hipStreamBeginCapture(graph_stream, hipStreamCaptureModeGlobal));
+    if(alg_mode == rocsolver_alg_mode_graph)
+    {
+        HIP_CHECK(hipStreamCreate(&graph_stream));
+        rocblas_set_stream(handle, graph_stream);
+        HIP_CHECK(hipStreamBeginCapture(graph_stream, hipStreamCaptureModeGlobal));
+    }
 
-#endif // ROCSOLVER_GEQRF_USE_HIPGRAPH
-
+    // execute computations
     while(j < dim - GEQxF_GEQx2_SWITCHSIZE)
     {
         // Factor diagonal and subdiagonal blocks
@@ -177,20 +179,19 @@ rocblas_status rocsolver_geqrf_template(rocblas_handle handle,
         j += nb;
     }
 
-#ifdef ROCSOLVER_GEQRF_USE_HIPGRAPH
+    if(alg_mode == rocsolver_alg_mode_graph)
+    {
+        hipGraph_t graph;
+        HIP_CHECK(hipStreamEndCapture(graph_stream, &graph));
+        rocblas_set_stream(handle, stream);
 
-    hipGraph_t graph;
-    HIP_CHECK(hipStreamEndCapture(graph_stream, &graph));
-    rocblas_set_stream(handle, stream);
-
-    hipGraphExec_t exec;
-    HIP_CHECK(hipGraphInstantiate(&exec, graph, nullptr, nullptr, 0));
-    HIP_CHECK(hipGraphDestroy(graph));
-    HIP_CHECK(hipGraphLaunch(exec, stream));
-    HIP_CHECK(hipGraphExecDestroy(exec));
-    HIP_CHECK(hipStreamDestroy(graph_stream));
-
-#endif // ROCSOLVER_GEQRF_USE_HIPGRAPH
+        hipGraphExec_t exec;
+        HIP_CHECK(hipGraphInstantiate(&exec, graph, nullptr, nullptr, 0));
+        HIP_CHECK(hipGraphDestroy(graph));
+        HIP_CHECK(hipGraphLaunch(exec, stream));
+        HIP_CHECK(hipGraphExecDestroy(exec));
+        HIP_CHECK(hipStreamDestroy(graph_stream));
+    }
 
     // factor last block
     if(j < dim)